refactor: centralize delegate recovery semantics
