+++ crates/app/src/conversation/turn_coordinator.rs
@@ -127,11 +127,8 @@ use super::turn_shared::{
 #[cfg(test)]
 use super::turn_shared::{ReplyResolutionMode, ToolDrivenFollowupKind};
 use crate::config::ToolConsentMode;
-#[cfg(feature = "memory-sqlite")]
-use crate::session::recovery::{
-    RECOVERY_EVENT_KIND, build_async_spawn_failure_recovery_payload,
-    build_terminal_finalize_recovery_payload,
-};
+#[cfg(all(feature = "memory-sqlite", test))]
+use crate::session::recovery::RECOVERY_EVENT_KIND;
 #[cfg(all(test, feature = "memory-sqlite"))]
 use crate::session::repository::TransitionApprovalRequestIfCurrentRequest;
 #[cfg(feature = "memory-sqlite")]
@@ -4344,7 +4341,7 @@ pub(crate) async fn run_started_delegate_child_turn_with_runtime<
     }
 }
 
-#[cfg(feature = "memory-sqlite")]
+#[cfg(all(feature = "memory-sqlite", test))]
 fn finalize_async_delegate_spawn_failure(
     memory_config: &MemoryRuntimeConfig,
     child_session_id: &str,
@@ -4353,37 +4350,14 @@ fn finalize_async_delegate_spawn_failure(
     execution: &ConstrainedSubagentExecution,
     error: String,
 ) -> Result<(), String> {
-    let repo = SessionRepository::new(memory_config)?;
-    let subagent_contract = execution.contract_view();
-    let outcome = crate::tools::delegate::delegate_error_outcome(
-        child_session_id.to_owned(),
-        Some(parent_session_id.to_owned()),
-        label,
-        Some(&subagent_contract),
-        error.clone(),
-        0,
-    );
-    let request = FinalizeSessionTerminalRequest {
-        state: SessionState::Failed,
-        last_error: Some(error.clone()),
-        event_kind: "delegate_spawn_failed".to_owned(),
-        actor_session_id: Some(parent_session_id.to_owned()),
-        event_payload_json: execution.terminal_payload(
-            ConstrainedSubagentTerminalReason::SpawnFailed,
-            0,
-            None,
-            Some(error.as_str()),
-        ),
-        outcome_status: outcome.status,
-        outcome_payload_json: outcome.payload,
-    };
-    finalize_terminal_if_current_allowing_stale_state(
-        &repo,
+    crate::operator::delegate_runtime::finalize_async_delegate_spawn_failure(
+        memory_config,
         child_session_id,
-        SessionState::Ready,
-        request,
-    )?;
-    Ok(())
+        parent_session_id,
+        label,
+        execution,
+        error,
+    )
 }
 
 #[cfg(feature = "memory-sqlite")]
@@ -4395,69 +4369,14 @@ fn finalize_async_delegate_spawn_failure_with_recovery(
     execution: &ConstrainedSubagentExecution,
     error: String,
 ) -> Result<(), String> {
-    let recovery_label = label.clone();
-    match finalize_async_delegate_spawn_failure(
+    crate::operator::delegate_runtime::finalize_async_delegate_spawn_failure_with_recovery(
         memory_config,
         child_session_id,
         parent_session_id,
         label,
         execution,
-        error.clone(),
-    ) {
-        Ok(()) => Ok(()),
-        Err(finalize_error) => {
-            let repo = SessionRepository::new(memory_config)?;
-            let recovery_error = format!(
-                "delegate_async_spawn_failure_persist_failed: {finalize_error}; original spawn error: {error}"
-            );
-            match repo.transition_session_with_event_if_current(
-                child_session_id,
-                TransitionSessionWithEventIfCurrentRequest {
-                    expected_state: SessionState::Ready,
-                    next_state: SessionState::Failed,
-                    last_error: Some(recovery_error.clone()),
-                    event_kind: RECOVERY_EVENT_KIND.to_owned(),
-                    actor_session_id: Some(parent_session_id.to_owned()),
-                    event_payload_json: build_async_spawn_failure_recovery_payload(
-                        recovery_label.as_deref(),
-                        &error,
-                        &recovery_error,
-                    ),
-                },
-            ) {
-                Ok(Some(_)) => Ok(()),
-                Ok(None) => {
-                    let current_state = repo
-                        .load_session(child_session_id)?
-                        .map(|session| session.state.as_str().to_owned())
-                        .unwrap_or_else(|| "missing".to_owned());
-                    Err(format!(
-                        "{recovery_error}; delegate_async_spawn_recovery_skipped_from_state: {current_state}"
-                    ))
-                }
-                Err(recovery_event_error) => match repo.update_session_state_if_current(
-                    child_session_id,
-                    SessionState::Ready,
-                    SessionState::Failed,
-                    Some(recovery_error.clone()),
-                ) {
-                    Ok(Some(_)) => Ok(()),
-                    Ok(None) => {
-                        let current_state = repo
-                            .load_session(child_session_id)?
-                            .map(|session| session.state.as_str().to_owned())
-                            .unwrap_or_else(|| "missing".to_owned());
-                        Err(format!(
-                            "{recovery_error}; delegate_async_spawn_recovery_skipped_from_state: {current_state}"
-                        ))
-                    }
-                    Err(mark_error) => Err(format!(
-                        "{recovery_error}; delegate_async_spawn_recovery_failed: {mark_error}; delegate_async_spawn_recovery_event_failed: {recovery_event_error}"
-                    )),
-                },
-            }
-        }
-    }
+        error,
+    )
 }
 
 #[cfg(feature = "memory-sqlite")]
@@ -4593,89 +4512,11 @@ fn finalize_delegate_child_terminal_with_recovery(
     child_session_id: &str,
     request: FinalizeSessionTerminalRequest,
 ) -> Result<(), String> {
-    let recovery_request = request.clone();
-    match finalize_terminal_if_current_allowing_stale_state(
+    crate::operator::delegate_runtime::finalize_delegate_child_terminal_with_recovery(
         repo,
         child_session_id,
-        SessionState::Running,
         request,
-    ) {
-        Ok(()) => Ok(()),
-        Err(finalize_error) => {
-            let recovery_error = format!("delegate_terminal_finalize_failed: {finalize_error}");
-            match repo.transition_session_with_event_if_current(
-                child_session_id,
-                TransitionSessionWithEventIfCurrentRequest {
-                    expected_state: SessionState::Running,
-                    next_state: SessionState::Failed,
-                    last_error: Some(recovery_error.clone()),
-                    event_kind: RECOVERY_EVENT_KIND.to_owned(),
-                    actor_session_id: recovery_request.actor_session_id.clone(),
-                    event_payload_json: build_terminal_finalize_recovery_payload(
-                        &recovery_request,
-                        &recovery_error,
-                    ),
-                },
-            ) {
-                Ok(Some(_)) => Err(recovery_error),
-                Ok(None) => {
-                    delegate_terminal_recovery_skipped_error(repo, child_session_id, recovery_error)
-                }
-                Err(recovery_event_error) => match repo.update_session_state_if_current(
-                    child_session_id,
-                    SessionState::Running,
-                    SessionState::Failed,
-                    Some(recovery_error.clone()),
-                ) {
-                    Ok(Some(_)) => Err(format!(
-                        "{recovery_error}; delegate_terminal_recovery_event_failed: {recovery_event_error}"
-                    )),
-                    Ok(None) => delegate_terminal_recovery_skipped_error(
-                        repo,
-                        child_session_id,
-                        recovery_error,
-                    ),
-                    Err(mark_error) => Err(format!(
-                        "{recovery_error}; delegate_terminal_recovery_failed: {mark_error}"
-                    )),
-                },
-            }
-        }
-    }
-}
-
-#[cfg(feature = "memory-sqlite")]
-fn finalize_terminal_if_current_allowing_stale_state(
-    repo: &SessionRepository,
-    session_id: &str,
-    expected_state: SessionState,
-    request: FinalizeSessionTerminalRequest,
-) -> Result<(), String> {
-    match repo.finalize_session_terminal_if_current(session_id, expected_state, request)? {
-        Some(_) => Ok(()),
-        None => {
-            if repo.load_session(session_id)?.is_some() {
-                Ok(())
-            } else {
-                Err(format!("session `{session_id}` not found"))
-            }
-        }
-    }
-}
-
-#[cfg(feature = "memory-sqlite")]
-fn delegate_terminal_recovery_skipped_error(
-    repo: &SessionRepository,
-    child_session_id: &str,
-    recovery_error: String,
-) -> Result<(), String> {
-    let current_state = repo
-        .load_session(child_session_id)?
-        .map(|session| session.state.as_str().to_owned())
-        .unwrap_or_else(|| "missing".to_owned());
-    Err(format!(
-        "{recovery_error}; delegate_terminal_recovery_skipped_from_state: {current_state}"
-    ))
+    )
 }
 
 #[cfg(feature = "memory-sqlite")]
